Updated global and main

+++ global.c
@@ -2,8 +2,9 @@
 #include "global.h"
 
 // Function to calculate time
-float time_diff(struct timeval *start, struct timeval *end)
+unsigned long time_diff(struct timeval *start, struct timeval *end)
 {
     return (end->tv_sec - start->tv_sec) * 1000000 + (end->tv_usec - start->tv_usec);
 }
+
