+++ global.h
@@ -2,6 +2,8 @@
 #define LAB1_GLOBAL_H
 
 #define MAX 65535
-float time_diff(struct timeval *start, struct timeval *end);
+
+#include <sys/time.h>
+unsigned long time_diff(struct timeval *start, struct timeval *end);
 
 #endif //LAB1_GLOBAL_H
